Added mandatory iOS 10 privacy description
Since iOS 6 Apple provided a way to show the user why a certain permission is required. In iOS 10 this becomes mandatory, so any app not specifying the key will crash at runtime when a permissions is requested.

This plugin relies requests the Contacts permission, so at the very least we need to add `NSContactsUsageDescription` to the `plist`.

Note that this is the same approach as the Geolocation plugin has taken a while ago.