Update CDVContacts.m
This update fixes the issue https://issues.apache.org/jira/browse/CB-11532
I am not an iOS dev so am sure the code can be improved as it seems a bit heavy to create new dictionaries for each user. 

I have tested this in the wild with a phone containing the contacts below and it has solved the issue. 

GÃ¶kmen
BrÃ»lÃ©
FranÃ§a
Ø§Ø¯ÙˆØ§Ø±Ø¯
Gabby â€¢â€¢â€¢â€¢
DamiÃ¡n M.A.